Python 3 bug fix for the N-state model target function setup.
The num_tensors variable needs to be an integer, but the Python 3 division will create a float type.

+++ maths_fns/n_state_model.py
@@ -191,7 +191,7 @@ class N_state_opt:
 
             # Tensor set up.
             self.full_tensors = array(full_tensors, float64)
-            self.num_tensors = len(self.full_tensors) / 5
+            self.num_tensors = int(len(self.full_tensors) / 5)
             self.red_data = red_data
             self.red_errors = red_errors
             self.full_in_ref_frame = full_in_ref_frame
